Change state space representation
Particles now have 3-component axis/angle representation
for the orientation of a 3D obbject instead of Euler RPY.
In particular, the last three components of each particles
are no longer the Roll, Pitch, Yaw angles, but represent
the rotational versor multiplied by the angle rotation.
As a result, the rotation angle can be reconstructed usign the norm
of the last three components of each particle.