Apply major code refactoring for 1st working impl
Several new member functions to HTParticleFilteringFunction class
have been added. In particular, a new state observation and
correction methods have been implemented. They are similar to the
original one, but differ in
 - observation: no longer perform Canny edge detection, but simply
   superimposes the CAD model of the hand onto the camera image
 - correction: crop the camera image with a ROI generated from the
   rendered hand model, then performs HOG on both images.
Finally, the prediction step is not performed on the 5 particles
which most represents the hand on the camera image, i.e. the 5
particles corresponding to the 5 highest weights.

Several //FIXME tags have been added to the code.
They will be taken into account in the upcoming commits.