Fix closing issue when pressing ESC
The program wasn't closing properly when EXC button was pressed
because the runFilter() method needed glfwPollEvents() invokation
to properly exits. As a result, the std::thread was unable to
handle this need and it was replaced by std::async.
std::async is returns a std::future object which provides the
wait_for() method to query whether the thread has finished its task
or not. By setting a reasonable time interval for the wait_for()
call, i.e. 1 ms, it is possible to poll the thread status while
invoking the glfwPollEvents() in a while() loop.