Add scaled (larger) forearm
The iCub mesh of forearm is of V1, which is thinner than the one with the skin patches.
A new mesh of the forearm has been added which takes into account the bigger dimensions.
It is worth pointing out that the mesh has been edited from the V1 using Blender.