Add threading capability to VisualPF correction
A brand new implementation of the VisualParticleFilterCorrection
has been implemented.

In particular, at runtime, a number of threads equal to
the number of CUDA streams used by HOG are deployed.
In this way, each thread invokes, in parallel, the HOG kernels
on a different CUDA stream.

Code is still under profiling. The current implementation does not
provide the level of concurrency required to have significant
benefits over a serial implementation.