Add CUDA stream capability to VisualPF correction
A brand new implementation of the VisualParticleFilterCorrection
has been implemented.

Several HOG kernels are invoked on different Streams in a way that
(partial) parallelization is possible.

At the current stage, it seems that downloading results from the GPU
(i.e. performing DtoH memcopy) blocks Stream parallelization.
This issue is under investigation.

It is worth pointing out that
    Mat::setDefaultAllocator(cuda::HostMem::getAllocator(
    cuda::HostMem::PAGE_LOCKED));
is commented.

Code is still under profiling.
Partial stream parallelization is detected.