Remove code to initialize OpenGL context
To account for the new SICAD implementation, the code for initializing
the OpenGL context has been removed.

SICAD exposes two methods for initializing the OpenGL context and to
check whether the OpenGL window should close or not.
Thus, hand-tracking main program no longer requires to deal with
OpenGL context and, instad, uses the SICAD static methods (exposed
by the VisualProprioception class) to control the status of
OpenGL context and window.