Add new implementation for motion models
This commit add a new implementation for the motion models.
In particular, a new motion model using the icub forward
kinematics is implemented as a decorator to the Brownian motion.

Furthermore, this commit takes into accounte the new implementation of
the BayesFiltersLib which provides set property method for
motion and observation models.