Add first implementation of pose visual servoing
When the angular velocity of the rotation to be done by the
robot to reach the desired orientation falls below 3.0 degrees per
second, then the translational velocities are enabled to also reach the
desired position.

Translational velocities are enabled smoothly with an exponential
control low that saturate at 1.0.

The approach is still rudimental but it's the first step toward
more sophisticated techniques.