Add gpu counter variable to balance gpu load
gpu_count can be used to balance the computational load on the GPU.
such a variable could be equal to
cuda::DeviceInfo::multiProcessorCount(), but this does not imply that
the performance will improve because of a balanced load.

CUDA is pretty smart in automatically balance the computational burden
on the GPU and in some configuration (e.g. with a K40 GPU), a
gpu_count = 1 maximizes the performance.