Add new image-based visual servo control
The new visual servo control is based on the paper
    E. Malis, “Improving vision-based control using efficient
    second-order minimization techniques”, IEEE ICRA, vol. 2,
    p. 1843–1848, 2004.
that is an image-based visual servoing control law which should be
robust aginst unwanted translational behaviour due to the presence of
orientation error.

The current implementation is a copy-pasted and modified version of
the existing one. In the upcoming releases, a new software architecture
will be devised to cope with possible different implementation of
visual servoing control techniques.