Implement VS with all feature pixel components
With this commit the visual servo control algoritm now have 16
pixel feature points instead of 12. Prior to this commit, we assumed
that the stereo cameras were at the same height so that to have the same
v coordinate pixel. Due to possible calibration or mechanical errors,
we drop this assumption and extend the control feature from 12 to 16,
i.e. 4 pixel coordinates from the left and right camera.